// types/ProductCategoryType.ts
export type ProductCategoryType =
  | "none"
  | "Books & Academic Materials"
  | "Electronics & Gadgets"
  | "Equipments"
  | "Fashion & Clothing"
  | "Hostel & Room Essentials"
  | "Kitchen & Food Items"
  | "Personal Care & Beauty"
  | "Sportswear"
  | "Stationery & Office Supplies"
  | "Transportation & Mobility";